Updating math functions in the c-library
Most of the math functions pass the tests, still have some issues with
pow, which some of the other functions rely on, and some of the
hyperbolic functions

+++ clib/src/stdio/stream.c
@@ -396,13 +396,13 @@ case_e:
 		*prefix = (" ");
 
 	/* Handle special cases first */
-	if (_isnan((double)fpval))
+	if (_isnan(fpval))
 	{
 		(*string) -= sizeof(_nan) / sizeof(char) - 1;
 		strcpy((*string), _nan);
 		fpval2 = 1;
 	}
-	else if (!_finite((double)fpval))
+	else if (!_finite(fpval))
 	{
 		(*string) -= sizeof(_infinity) / sizeof(char) - 1;
 		strcpy((*string), _infinity);
